refactor: Modal 관련 타입 분리
